add submodule PHPMailer and fix&refactoring

+++ urls-cron.php
@@ -1,6 +1,6 @@
 <?php
   require_once(dirname(__file__) . '/urls-config.php');
-  require_once(dirname(__file__) . '/urls-lib/core.php');
+  require(dirname(__file__) . '/urls-lib/core.php');
   
   $now = date('Y-m-d H:m:s');
 
@@ -20,25 +20,25 @@
     if($daysago > 14) {
       echo $daysago . "\n";
       $result = $db->delete_shorturl($lid);
-      var_dump($result);
     }
   }
   
   $tokenstatus = $db->get_token();
   $tokens = $tokenstatus['rows'];
   $tokendatas = $tokenstatus['data'];
+  
+  if($tokens > 0) {
+    foreach(range(0, $tokens - 1) as $rownum) {
+      $tokendata = $tokendatas[$rownum];
 
-  foreach(range(0, $tokens - 1) as $rownum) {
-    $tokendata = $tokendatas[$rownum];
+      $token = $tokendata['token'];
+      $createdate = $tokendata['create_date'];
+      $daysago = (strtotime($now) - strtotime($createdate)) / (60*60);
 
-    $token = $tokendata['token'];
-    $createdate = $tokendata['create_date'];
-    $daysago = (strtotime($now) - strtotime($createdate)) / (60*60);
-
-    if($daysago > 24) {
-      echo $daysago . "\n";
-      $result = $db->remove_token($token);
-      var_dump($result);
+      if($daysago > 24) {
+        echo $daysago . "\n";
+        $result = $db->remove_token($token);
+      }
     }
   }
